Speed: Double =>Float

+++ backend/src/main/scala/com/neo/sk/gypsy/gypsyServer/GameServer.scala
@@ -490,7 +490,7 @@ class GameServer(override val boundary: Point) extends Grid {
             hasMoved =true
             newMass = (newMass + p.mass).toShort
             newRadius = Mass2Radius(newMass)
-            newSpeed = sqrt(pow(vx,2)+ pow(vy,2))
+            newSpeed = sqrt(pow(vx,2)+ pow(vy,2)).toFloat
             newTargetX = vx.toShort
             newTargetY = vy.toShort
             massList = massList.filterNot(l => l == p)
